backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
asyncpg>=0.28.0
python-multipart>=0.0.6
email-validator>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
httpx>=0.25.0
passlib>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
